Define both column contents in each view's file
